app/requirements.txt
asgiref==3.4.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cymem==2.0.6
Cython==0.29.32
dataclasses==0.8
Django==3.2.15
django-filter==21.1
djangorestframework==3.13.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
ez-setup==0.9
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.4.0/fr_core_news_sm-3.4.0-py3-none-any.whl
idna==3.3
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.0.1
murmurhash==1.0.7
numpy==1.19.5
packaging==21.3
pathy==0.6.2
pkg_resources==0.0.0
preshed==3.0.6
pydantic==1.9.1
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.9
spacy-loggers==1.0.3
sqlparse==0.4.2
srsly==2.4.4
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.1.1
urllib3==1.26.11
wasabi==0.10.1
zipp==3.6.0
